themes: Correct Gruvbox color codes.
Missing color codes for `dark cyan` and `brown` were added using
https://github.com/morhetz/gruvbox/blob/master/colors/gruvbox.vim
as reference.

`popup_category` style did not include BOLD for the 256 and 24bit
versions while 16 color did. Hence, BOLD has been added for the
other 2 as well.

`starred_count` and `footer` style did not have a `light gray`
option hence GRAY which was actually DARKGRAY was used previously.
Now, an appropriate LIGHTGRAY has been added.

To do so Light2 was used for WHITE and Light3 was used for LIGHTGRAY.