bugfix: boxes/model/themes/ui: Add & apply footer_contrast style.
This commit adds entries in themes.py, labelled 'footer_contrast' that
define the color for hint text. This could also be referred to as the
color scheme for background of the footer.

Prior to this an old theme style name was in use, which could give
unreliable rendering results between systems. This commit migrates from
this old absent "code" style to use the newly added style.

Tests amended.