refactor: model/views/test_ui_tools: Import submodule from `datetime`.
This commit imports the particular submodule from datetime instead
of importing only the main datetime library in model.py, views.py and
test_ui_tools.py.

This allows for example in model.py, to shorten `datetime.datetime` to
`datetime` and allows for easier import and use of other modules like
`date` or `timedelta`.

This also improves consistency with `boxes.py`.