tests: Remove unsupported `short_name` field.
The unused `short_name` field is completely removed from the codebase
after the merge of zulip/zulip#15831 to the server repo that replaces
the short_name field with full_name in user profiles.

According to the API changelog, the removal was done in phases, where
in the first phase the `short_name` field was removed from GET /messages
endpoint in server version 3.1 / ZFL 26. In the second phase the
field was removed from POST /users endpoint in server version 3.1 and
ZFL 27.