api_types/helper: Remove deprecated `recipient_id` field.
This commit removes the deprecated Message `recipient_id` field from the
entire codebase and adds a comment in api_types to indicate that the
field is no longer used/needed. The server hasn't removed the field yet,
but it recommends clients to use the `display_recipient` field to get
info about the list of recipient(s).

Fixtures & tests amended.