bugfix: boxes: Set propagate_mode to change_one if topic unchanged.
The intent of the edit/update message API has recently been clarified
that if the topic is unchanged (even if sent) that the propagate_mode
should be set to change_one. This was previously not enforced by the
server but was changed to do so in
https://github.com/zulip/zulip/commit/e231a03eff6cce62243284ebcf7f06f798715a78

Since our propagate_mode defaults to change_later via the UI element,
the above change to the server leads to errors upon attempting
content-only edits.

To resolve this, the original topic is added to the new tuple form of
msg_edit_state, allowing detection of whether the topic has changed and
setting of propagate_mode as appropriate.

The decision to use change_one is dependent upon the trimmed new topic
since otherwise the edit can confusingly fail since trimming is checked
on the server end and results in the same propagate error message.

Tests updated.