boxes: Use an indicator attribute to specify current compose box status.
This commit introduces an attribute, `compose_box_status` to the WriteBox
class, to indicate the current compose status of the WriteBox, which can
be one of "open_with_private", "open_with_stream" or "closed".

Such an indicator is a structured way to check for compose status before
performing status specific tasks, and is much clearer to the reader than
testing the state of associated attributes (eg. to_write_box, stream_id)

Tests amended.