color: Add default color-scheme with bold, based on existing colors.
A DefaultColor class is added to color.py which acts as the default
color-scheme for ZT. The BOLD property is added to these colors
using the helper function color_properties.

The format:
* The colors have been defined using Enum's as a single string
containing 16, 256 and 24 bit codes.
* Each style is defined as a tuple of 2 elements, the foreground
and the background.

The 24 bit colors are not defined for the default colors hence
default is used. 2 missing colors in DEF, DARK_MAGENTA and
LIGHT_CYAN, which were used in zt_blue were added into Color.