platform_code/helper: Create new module to hold platform dependent code.
This commit creates a new module (platform_code.py) inside zulipterminal/
which contains the code to detect the current platform (OS) that the ZT
instance is running on.

The platform_code.py module has a PLATFORM variable that holds the platform
name which could only be one of "Linux", "MacOS", "WSL" or "unsupported".
Currently, we only use this to set our old platform variables in helper.py
accordingly.