refactor: boxes/model: Verify each email individually through model.
Previously, a list of emails were expected as a parameter in the
model helper `get_invalid_recipient_emails` which has been refactored
to `is_valid_private_recipient` to abstract the processing of the list
to `boxes.py` and rely on the model helper to verify each recipient
individually.

The model helper checks whether a given recipient email exists in the
`user_dict` and returns the relevant boolean.