refactor: boxes: Extract current recipient validation into a helper.
This commit extracts the current private message recipient validation
into a helper method, `_validate_recipients_and_notify_invalid_ones`,
that returns a bool, indicating whether or not, all of the recipient
emails mentioned in the `to_write_box` are valid (returns False even on
a single invalid email). The `CYCLE_COMPOSE_FOCUS` keypress now relies
on this helper method to specify whether or not to continue with it's
flow.