refactor: platform_code/helper/model: Migrate notify() to platform_code.
This commit shifts the notify() function to the platform_code.py module
in an effort to move all platform dependent code to a single isolated
module so that other modules are unaware of the underlying platform and
how they are handled for different platforms. This will also allow us to
better handle unsupported platforms.

Tests amended.

New test file created, and added to list of files to be linted by mypy.