refactor: tests: ui_tools: Inline EDIT_MESSAGE keypress test ids.
This makes it clearer which test case has which id, compared to having
separate lists of test case values and ids, which could get difficult
to compare, especially when the lists get long.

This will potentially be helpful when we add more tests for handling
all range of message editing possibilities.