refactor: tests: ui_tools: Move and rename parameters of EDIT_MESSAGE.
This renames the "msg_body_edit_enabled" parameter of EDIT_MESSAGE keypress
to an expected type, since its not an input. Also, moves it to the last of
the parameters where we generally put expect_*, to maintain consistency.

We also move the expect_editing_to_succeed parameter to the end to
maintain consistency.