refactor: tests: ui_tools: Reduce duplication in mouse scroll tests.
This adds a fixture that generates the required parameters for mouse scroll
actions. We return the type of event i.e. mouse press, along with the button
and the expected key to be triggered for this mouse_event.

Introducing the mouse_scroll_event fixture in mouse_event tests for
several *Views classes reduces code duplication - essentially extracting
duplicated test code and parametrizes into one fixture, making the tests
more compact and readable.

Co-authored-by: Preet Mishra <ipreetmishra@gmail.com>