config/themes: Add "widget_disabled" style attribute for all themes.
This commit adds a "widget_disabled" style for all themes that could be
passed as a Display attribute to any widget which needs to be disabled
for certain cases. This atttribute can be coupled with urwid.WidgetDisable
to effectively produce a completely disabled widget both visually and
functionally.

This renders the widget foreground in DARK_GRAY in most themes and the
background is kept as the default theme background. For 1-bit color-depth
the widget will render with a strikethrough.