refactor: tests: core/conftest: Add & use index_search_messages fixture.
The `index_search_messages` fixture returns an initial index with the
"search" key of the index set with a `set` containing a single
`message_id`, 500.

The fixture is then used to specify the model's index in
`test_search_message()`, as opposed to the previous dict consisting of
only the "search" key.