tools: Include `test_run.py` under files to be type-checked with mypy.
This commit adds a new "tests" key to the `repo_python_files` dict to
let mypy know which files to type check. A new list,
`type_consistent_testfiles` is also added, to hold the list of test
files that have been type-annotated incrementally till all the test
files are consistent.