tools: Edit convert-unicode-emoji script to store extended emoji format.
This commit updates the convert-unicode-emoji-data script to store all the
unicode emoji codepoints in their extended format as received from the
server.

This script will map each emoji_name (aka canonical_name) to its data
containing the emoji_code and aliases that share the same emoji_code.
We sort this dictionary in ascending order of key (emoji_name) and store
it as an OrderedDict to maintain the order of sorting.

We also turn off black formatting by wrapping the file with fmt: off/on to
disable black from formatting the dictionaries afterwards - since that is
a generated file, it should not be modified.