unicode_emojis/helper/model: Re-generate EMOJI_DATA and update model.
This commit generates the EMOJI_DATA from tools/convert-unicode-emoji-data
in the new format that maps each emoji_name to its corresponding emoji's
code, and aliases that share the same emoji_code, stored in an OrderedDict.
The EMOJI_DATA is sorted in ascending order of emoji_name and this ordering
is maintained via OrderedDict. The generated EMOJI_DATA is stored in
unicode_emojis.py file.

This also updates the helper function in model that generates all emoji
data to follow and adapt to this new format of storing emoji's and updates
the EmojiData type to indicate each emoji now includes an aliases field.
This also migrates the emoji fixtures in conftest i.e. realm_emoji_data,
unicode_emojis and zulip_emoji to the new format of storing emoji_data.

Tests amended.