refactor: buttons/core: Avoid using StreamButton as a popup parameter.
Prior to this commit, the `stream_muting_confirmation_popup` required
a `StreamButton` instance as a parameter to mute/unmute it and show
it to the user to request for confirmation. This commit refactors
the parameters to require the `stream_name` and `stream_id` instead
to abstract the StreamButton from this method.

Tests updated.