bugfix: boxes/ui: Display mouse selection_hint for entire `View`.
This commit fixes 2 bugs related to displaying selection_hint on mouse_drag
events. This migrates the mouse_event from MessageBox to the entire View
so that the selection_hint is displayed even when the user tries to perform
mouse_drag events outside of MessageBox i.e. anywhere in View.

This also fixes selection_hint getting stuck in the footer if the mouse is
dragged in the MessageView but released outside of it.