refactor: Reformat tools/ with black, except quotes.
NOTE: tools/lister.py was imported from zulip/zulip as a helper for
tools/run-mypy and has not been edited directly significantly
subsequently.

As such, no specific allowances are made in applying black to lister.py,
since black preserves functionality and in future we expect to move to
the lister functionality present in zulint.