run/ui: Add `dynamic` layout option for variable terminal support.
A dynamic option is added it ZT which adapts to various terminal
widths to make it more readable. To shift from using a bool to
options `self.autohide` to changed to `self.layout`.

This makes ZT behave like autohide when in a small terminal which is
defined using MINNORMALWIDTH. After crossing this limit ZT behaves
like no-autohide until MAXNORMALWIDTH after which, side panel widths
are not fixed anymore and they start occupying space in the ratio
20:60:20. This gives the effect that the side panels are expanding.
This stops when we reach the max width set for ZT.

To do so this makes use of urwid's render method to get maxcols, a
Literal `self.mode` which takes values "small", "normal" and "wide"
and existing `show_left_panel` and `show_right_panel` methods with
small changes.