refactor: tests: conftest: Return a `Message` from template factory.
This commit changes the returned object of the `msg_template_factory`
from a `Dict` to a `Message` instance, which is the appropriate type
for a message template. `stream_id` is also included in the `Message`
instance only when it is a stream message.

Places where the factory function is used are also typed accordingly.