ui: Add TabView's to autohide layout.
This commit adds a closed tab view of width 3 for each of the side
panels when in the autohide layout. The initial layout when the app
is started is left panel open, right panel closed.

This makes use of the TabView class and changes show_left_panel
and show_right_panel accordingly.

Tests updated.