ui: Make side panels overlay the center panel in autohide layout.
This commit makes the side panels overlay the body which avoids
squashing and stretching of the message view, which improves UX in
autohide mode.

To do this the `show_panel` functions have been updated to use
Overlay.

The focus setting in the `show_panel` functions which could be
avoided at a later point needed a reordering in the SEARCH keypress
which actually makes it the correct order, i.e.,
* first handle autohide
* then handle search

Tests updated.