bugfix: model/core: End typing notification on unresponsive event.
The corner case of not receiving a stop signal is handled by
setting TYPING_STARTED_EXPIRY_PERIOD=15 and checking this
condition using datetime and comparing it with
`typing_start_time` state of `active_conversation_info`.

`typing_start_time` is used so that we don't miss a start
event (which is received every 10 seconds) and misinterpret it as
stop not received.

`active_conversation_info` is stored in a variable so that it
doesn't change dynamically while inside the loop.

Tests added.