core: Implement new asynchronous Typing Notification function.
A new asynchronous function `show_typing_notification` is introduced,
which is intended to initiate on 'start' events and be externally ceased
on 'stop' events through resetting an `active_conversation_info`
dictionary. The state in this dictionary controls the output while the
function is running, initially only the `sender_name`.

To ensure only one copy of the async function runs at once, a boolean
`is_typing_notification_in_progress` is used and can be checked at
calling points.

Tests added.