refactor: model: Migrate to new asynchronous Typing Notifications.
This commit refactors `_handle_typing_event` to use the new asynchronous
approach to typing notifications method in the Controller.

As per the description in the commit which introduced it, on a start
event we now:
* update the `sender_name` in the `active_conversation_info` state dict
* run the `show_typing_notification` method
  (if not already running - checking `is_typing_notification_in_progress`)

On a stop event we reset the state dict, which generally causes the
notification method to cease running.

Tests added.