refactor: ui/views/buttons: Combine autohide handling with focus.
This commit combines the focus setting with autohide handling. This
makes it easier to handle focus change of panels since both these
functions are often seen used together.

This also avoids the situation of the show_panels handling focus as
moving to a different panel through focus_panel takes care of that.

This also gives the advantage of knowing which panel we are on even
if we are in autohide layout hence giving us the ability to
manipulate the current panel (Eg: resizing) depending where we are.

Tests added.