bugfix/refactor: Fix 16color stream colors and their format.
This commit adds yellow as the stream color for all 16 color themes
instead of using the default and causing problems with different
colored/transparent terminals and also disturbing the focus
highlighting.

This also changes the format of stream color styles to follow the
same format as other styles generated in `parse_themefile` in
`themes.py` so that it becomes easier to manipulate them if
required.

Tests added.