refactor: api_types/helper/views: Type user's bot_type as Optional.
This commit modifies the `bot_type` attribute of the RealmUser class
to an `Optional[int]` from the previous `int` data type. This change
is motivated by the server not assigning an integer to the user when
the user isn't a bot, making None the ideal value in that case.

To maintain consistency the `bot_type` attribute of the TidiedUserInfo
class becomes an `Optional[int]` from the previous `int` data type.