bugfix: tests: boxes: Assert on to_write_box after private_box_view.
This commit adds assert statements within multiple tests to assert
that the value of WriteBox's `to_write_box` is None after a
`private_box_view` function call, since it initiates `to_write_box`
with a ReadlineEdit instance.

This is also necessary for supporting mypy, since the attribute type
depends on whether it is a private box or not.