buttons/api_types/ui_sizes: Indent to distinguish resolved topics.
This commit indents all TopicButtons to make space for the resolved
topic symbol and easily distinguish between them.

A RESOLVED_TOPIC symbol has been added to api_types to be used for
recognizing a resolved topic.

To make topic name length the same as before, LEFT_WIDTH is
increased by 2. This is equivalent to no difference in topic name
length and 2 char gain in stream/special narrow buttons.

Tests updated.