bugfix: model: Add updated `message_id`s to their new topic's set.
Prior to this commit, if a message's topic was changed, the
corresponding `message_id` was not added to the new topic's set
under the "topic_msg_ids" key of the model's index. This commit
adds the updated `message_id` to that set, ensuring the
accuracy of the new topic's set.

Tests updated.