feature: Add presence markers to PM compose.
Fixes part of #1155.

PMs didn't have a way to know whether the user was active without presence markers from the right column view, which is a hassle. This allows them to see it directly.

Need to add colour to the markers.

In the future, maybe add some form of last seen text that's shorter than just the timestamp ("last seen 3 hours ago") instead of ("last seen on Fri 3 Aug 03-02-2002").