keys: Narrow types required in KeyBinding using NotRequired.
Use of NotRequired is much more specific than use of total=False in a
TypedDict, so catches more incorrect uses.

NotRequired could be applied elsewhere in the code, but eg. Message is much
more complex due to it being part of an API which has changed over time.

Tests updated.