bugfix: core: Do not exit editor mode when narrowing.
These unrelated aspects being tied together was not an issue previously,
when narrowing was always done outside of compose. However, with
integration of #1194 as 2195904dd this is no longer the case, so this is
required to ensure that narrowing does not prematurely end editor mode,
since otherwise container widgets would steal keypresses.

This may be improved later through further refactoring of the keypress
handling.