tools: Add check-branch to run 'make check' on all branch commits.
This git command can be run manually, but is encoded here in a
reproducible format via this script, supporting use by maintainers and
new developers alike.

The branch for comparison is defined by default as from the merge-base
with upstream/main, though another comparison point can be specified as
the first parameter, eg. HEAD~2 to check the last two commits, or main
to specify the local main in case it differs from upstream/main.

The `git --no-pager log -1` ensures it is clear which commit is being
tested at each stage.

Check taken from tools/fetch-pull-request to handle situation with
uncommitted changes.

Other tools updated to exclude testing this bash script.