check-branch: Do nothing on empty branches & check for ambiguous main.
If current branch has no differences from main, skip check.
This would avoid checking a branch which simply makes changes and
reverts them, but this should be a rare case, and otherwise avoids
accidental extra testing of eg. main.

By default the script uses upstream/main as the merge-base; in case the
local main has differences, warn of this and require explicit
specification of the merge-base as the first parameter in order to run
the tool.