helper.py: Helper function to make dates user-friendly.
This commit adds a function that takes in a datetime string,
checks whether it came from message_info or user_info depending
on string length (later, we could include a boolean in the
arguments. For now, that's an unnecessary variable), and then
returns a user-friendly version of the date and time.

For example, if the string is
"Fri Apr 29 2022 05:15:12 AM", then converted will be something like
"A week ago", "43 minutes ago", and so on.