refactor: button: Simplify EmojiButton check mark update method.
EmojiButton currently relies upon calls in multiple locations to update_widget,
with a parameter derived from a helper method, in order to set the check-mark
status.

This commit migrates the style closer to update_count, with a new method name
update_check_mark, with the same functionality as the previous method except
for also calling update_widget internally.

This is clearer, reduces duplication and is less prone to error.

Tests updated.