refactor: test_boxes: Migrate to_pin to descriptive stream_categories.
In test_generic_autocomlete_streams, we change test parameter to_pin to
stream_categories.

For no pinned streams this new parameter should be an empty dictionary, and
initially we only support an optional "pinned" key which takes values identical
to those previously used, ie. a List[str]. This makes it clear that particular
streams are pinned when reading individual test cases, which could previously
be confusing given the large number of test cases.

It is intended to later use this parameter with other stream properties which
may affect ordering, such as for muted streams, which will further increase the
motivation to moving towards this design.