refactor: test_boxes: Assert all output strings match required_typeaheads.
This improves diagnosis of which states may be failing in test cases in
test_generic_autocomplete_streams.

When an assert is in a loop and one fails, others are not tested, which doesn't
always present a clear picture of what is wrong. For example, it could be that
multiple results are in the wrong order, and the code needs to generate them in
a different order to pass the test.