refactor: model/views: Migrate get_next_unread_topic into model.
While functional in the MiddleColumnView, it only relies upon state stored
there and only accessed by itself, as well as model data, so this migrates it
into the model along with that state. This will make it easier to test the core
functionality independently of the UI when we adjust the feature to depend upon
other existing model state such as muted streams and topics.

Also ensure the current behavior is maintained for now by explicitly sorting
the unread topics keys, which previous relied upon the initial and subsequent
data being ordered.

Tests migrated, simplified and set to use valid stream+topic data.