api_types: Add RealmUserEvent and RealmUserEventPerson.
This commit contains types for RealmUserEvent and RealmUserEventPerson added to
api_types.

Like a few other fields, custom_profile_field data may be sent and will be
updated for the user. It is commented out since we will want to implement this,
and we want to ensure this is done correctly, and may need a special case.