model: Add function to handle realm_user update events.
This commit contains Model._handle_realm_user_event() which updates the meta
data of the respective user in Model.initial_data["realm_users"].

A test is added for this in test_model.

Function and test logic slightly adjusted by neiljp to support differing event
and user data fields, as is necessary for new_email/email, without adjusting
the original API data-structure.

Note that after such an event, we currently rely upon the regular update of
users to propagate most of the new data around the application.