pyproject/setup/makefile/lint-all/lint-and-test: Replace flake8 by ruff.
This replaces flake8 with ruff for linting, while the fixing features of ruff
also replace autopep8 and autoflake.

The new tool, ruff, explicitly takes into account autoflake, and autopep8 does
not appear to currently leave code as-is with black, so is also removed here.

The initial ruff config is now in pyproject.toml, migrated automatically using
the flake8-to-ruff tool, then fine tuned to remove old plugins which are
of low value or unused (continuations, quotes).

On a local machine, flake8 took >2s per run, whereas ruff takes 1-2% of that time.
This is fairly negligible, but improves responsiveness - and this is before
enabling more checks that ruff offers.