setup/README: Clarify Python version support is 3.7-3.11 (remove 3.6).
Security support for Python 3.6 ended in December 2021.

Testing against 3.6 was already removed in CI in de2caaf5d and 926fc4c96.
Testing against 3.11 was already added in CI in 8c61bd44d.

Adjust python_requires range to disallow installation on Python 3.6; 3.11 was
already possible due to the change mentioned above to support testing in CI.

README updated to indicate the last release with Python 3.6 support was 0.7.0,
and later will require Python 3.7, but that we now also test against 3.11.

Package classifiers adjusted to add Python 3.11, and Python 3 updated to be 3-only.

Fixes #984.