FAQ: Expand upon & split previous truncated-link subsection.
The 'Links don't render completely as a footlink' (sub)section attempted to
document that footlinks were truncated intentionally.

The elements of that specifically related to footlink layout are now in a
separate subsection. An additional subsection then follows to indicate how
to access links which may have been partially hidden as footlinks due to
truncation or the maximum-footlink setting.