README: Improve documentation of configuration file.
The introduction to this section is reworded to explicitly state that the api
section can be automatically generated, while the zterm section is added
manually.

The example configuration file is adjusted:
- uncommented settings now relate to default values
- the incompatible footlinks and maximum-footlinks settings are not both uncommented
- documentation of setting names/meanings/values should now be clearer
- double-commenting is now used for setting notes, to avoid conflict with commented values
- extra lines are used to make settings more readable, and group footlinks settings

These changes lead to a zterm configuration section that can be dropped into a
zuliprc file and work as if the default values were used, but with improved
documentation/layout.

This should help towards avoiding issues such as #1269.