docs/buttons: Expand PMs to Private Messages in documentation.
Using abbreviations can be confusing, particularly for new users, and is
ideally avoided during most discussions and documentation.

In addition, an abbreviated form is not used for stream-messages, so it is also
clearer if both are written out in full.

This does not update any similar abbreviations in the code, except in the
docstring of buttons.py, which is used to autogenerate the
developer-file-overview.md file.

NOTE: This commit is in preparation for an intended upcoming change to use
'Direct Messages' for this naming in #1288.