lint-and-test: Run check-branch between basic checks & multi-platform.
This runs tools/check-branch as an additional job between the group of basic
linting and tests jobs, and the matrix of jobs testing on many platforms, for
pull requests only. For other events, the workflow skips all steps in the job,
causing the job to succeed and the workflow to proceed to run the matrix of
tests.

This does increase the total test time required, if all jobs run to completion,
which scales up with the number of commits, currently ~ 30s + 50s/commit.
However, with three groups of dependent jobs, work in progress PRs are likely
to not run all groups and instead stop midway, meaning contributor workflow
should not be sigificantly affected. In addition, this can likely be optimized
further, since the current duration includes an additional initial install, as
well as running checks on the final commit, which are effectively already
tested elsewhere.

Enforcing isolated commits in CI in this way has multiple benefits:
- Documentation can indicate that isolated commits are preferred, but requires
  that contributors are aware of it
- Developer bandwidth is limited in referring contributors to documentation
- Developer time in also limited to locally test that PRs meet this condition
- Contributors can identify what the problem is via a CI job failure

Note that isolated commits are not the same as being coherent:
- CI is only checking that each commit individually passes linting and tests
- there is no check that extra tests, error handling, etc. are added in the commit
(ref https://zulip.readthedocs.io/en/latest/contributing/commit-discipline.html)

However, to pass this job:
- existing tests should need to be adapted when code is changed
- commits will not get their linting/test status fixed only in later commits