requirements[dev]: Upgrade mypy from ==0.961 to ~=1.0.0.
This is necessary since fresh environment builds (including CI) appear to pull
a version of types-requests that is broken with mypy < 1.0, types-requests
being the broken-out 3rd-party package of stubs for requests, but that is still
maintained in typeshed.

This appears to be python/typeshed#9702.

The mypy version is updated to ~=1.0.0 versioning now that it has moved to a
major/minor/point versioning scheme.