setup: Remove deprecated 'setup.py test' command.
This was added in the first version of setup.py for the project. Since then,
setuptools has focused more on becoming a build tool (see eg.
pypa/setuptools#1684), and the project has developed a more refined testing
infrastructure which doesn't explicitly use this functionality.

If remaining, this code would require a noqa comment for compliance with the
incoming `N` set of linting rules, which is now no longer necessary.