requirements[dev]: Upgrade gitlint from  >=0.17 to ~=0.18.0.
This is possible given 0.18.0 dropped Python 3.6 support, and we have now
upgraded to a minimum of 3.7. However, this pins at ~=0.18.0 to avoid
accidentally losing support for Python 3.7, since the versioning scheme is
unclear and this tool is sufficiently useful as it stands.

This also sets an option to disable a warning regarding a regex that gitlint
was giving regarding titles, which the rule we use should be fine with.