bugfix: lint-and-test: Ensure check-branch uses dependencies per-commit.
Previously the implicit use of `make check` at each commit appears to update
the installed dependencies correctly per-commit - but this virtual environment
was not activated.

Instead the dependencies used were those from the previous job step, which were
those corresponding to the state at the end of the PR. This caused problems
such as
https://github.com/zulip/zulip-terminal/actions/runs/4290638353/jobs/7475275319,
where intermediate commits changed dependencies and updated files accordingly,
but different to the state at the final commit.

This commit migrates the dependency installation into the per-commit command,
now possible after the previous commit, combined with explicit linting and
tests to run on each commit (previously implicit as per `make check`).