refactor: run: Use helper function to simplify output of settings.
The output of each setting is indented and takes a standard form, so this makes
it clearer and easier to adjust.

The function takes a third parameter which is currently of use for clarifying
extra source data beyond the standard strings, such as aliased themes and
footlinks vs maximum_footlinks.