README: Add Structuring Commits section to Contributor Guidelines.
This documentation has long been needed in this repo, though has been covered
in Zulip documentation in some detail.

The text aims to lead between the concepts of developmental-style commits to
that of minimal coherent commits, motivated by effectively communicating code
changes and respecting reviewers' time, as well as emphasizing the various
benefits of the minimal coherent commit approach.