run/README: Use 'Zulip Server URL' uniformly.
This change was prompted by a user being unsure what a 'Zulip URL'
represents in the prompt for a username and password by the application,
if the user doesn't have an existing zuliprc file.

Including 'Zulip' and 'Server' makes this clearer, and we here take the
opportunity to synchronize this with a reference to something similar in the
README.

The mobile app also appears to use this terminology.