bugfix: run: Respect the server's choice of URL when fetching api key.
In the case where we connect to the server to fetch the api_key, but don't use
the preferred realm_url (with protocol), the server gives a 405 error. This is
interpreted by the current logic as that the email or password is incorrect,
which is confusing to a user.

Instead, use the server data that we already have available to adjust the
entered value to that preferred by the server.

The preferred URL is then passed back to be written in any created zuliprc.

Test updated.

This code could benefit from additional error handling, but this fixes #1318.