run-spellcheck/pyproject: Run both codespell and typos.
`typos` supports a toml file, but not pyproject.toml yet (see
crate-ci/typos#361). To avoid yet-another-config-file, this adds a list of
files to exclude in run-spellcheck, passed to typos on the command-line.

`codespell` excludes are already configured in pyproject.toml.

Since file excludes and reasons are added to run-spellcheck, explicitly
exclude it from being checked for both tools.