setup/pyproject.toml: Migrate fully from setup.py to pyproject.toml.
This transfers details from setup.py to equivalent ones in pyproject.toml,
and similarly removes or substitutes references to these files.

Note: This introduces two changes from the previous state:
* Building packages no longer removes the top title line shown on GitHub
* The `zip_safe` property is detected automatically by `setuptools`

This also adds a note to the making-a-new-release doc, since while it was
previously out of date, setup.py will now be absent completely.