MANIFEST/setup: Use setup.py & new MANIFEST.in to minimize package size.
Previously setup.py specifically excluded tests from being packaged, but recent
testing demonstrated this now included other files, including development
tools.

Moving to use the `include` option for the source is simpler here, combined
with the addition of a MANIFEST.in file.

The MANIFEST.in file is used by various tools, and appears more reliable than
other methods to achieve this, particularly as we move towards using
pyproject.toml.

For now, this aims to remove all but the absolute minimum files to run.