model: Support notification parameters to update_stream_message.
These parameters control the sending of notification messages upon moving
messages around, equivalent to renaming message topics.

Until now this code relied upon the server defaults, since the feature on the
server was introduced, and did not affect renaming topics within a stream.
However, this changed in ZFL 152 (Zulip 6) to also affect moving topics within
a stream, which differed from the previous behavior, leading to notification
messages upon renaming topics when parameters were not passed to the server.

To regain the older behavior, this method now explicitly passes these
parameters to the server (when they are supported), and uses default values
internally which are backwards compatible to how ZT used to behave, ie. not
sending notifications by default.

However, the existence of these parameters in the code will enable these
features to be used in future via this method.

For API documentation see
https://zulip.com/api/update-message#parameter-send_notification_to_old_thread

Test added.