README: Add Section welcoming users & developers to chat on czo.
This expands upon the existing welcoming text to discuss how one can
interact with those on chat.zulip.org and the stream, and expectations
of participants.

In particular, this pulls out some community norms that are more applicable,
and relates those to hotkeys in zulip-term.

Feature status section adjusted to link to new section.